requirements.txt
Django>=5.0,<6.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
whitenoise>=6.5.0
dj-database-url>=2.0.0
python-decouple>=3.8
boto3>=1.26.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
